Use doubles instead of float for min/max layer scales

+++ tests/src/core/testqgsmaplayer.cpp
@@ -123,24 +123,24 @@ void TestQgsMapLayer::setBlendMode()
 
 void TestQgsMapLayer::isInScaleRange_data()
 {
-  QTest::addColumn<float>( "scale" );
+  QTest::addColumn<double>( "scale" );
   QTest::addColumn<bool>( "isInScale" );
 
-  QTest::newRow( "in the middle" ) << 3000.0f << true;
-  QTest::newRow( "too low" ) << 1000.0f << false;
-  QTest::newRow( "too high" ) << 6000.0f << false;
-  QTest::newRow( "max is not inclusive" ) << 5000.0f << false;
-  QTest::newRow( "min is inclusive" ) << 2500.0f << true;
-  QTest::newRow( "min is inclusive even with conversion errors" ) << static_cast< float >( 1.0f / (( float )1.0 / 2500.0 ) ) << true;
+  QTest::newRow( "in the middle" ) << 3000.0 << true;
+  QTest::newRow( "too low" ) << 1000.0 << false;
+  QTest::newRow( "too high" ) << 6000.0 << false;
+  QTest::newRow( "max is not inclusive" ) << 5000.0 << false;
+  QTest::newRow( "min is inclusive" ) << 2500.0 << true;
+  QTest::newRow( "min is inclusive even with conversion errors" ) << static_cast< double >( 1.0f / (( float )1.0 / 2500.0 ) ) << true;
 }
 
 void TestQgsMapLayer::isInScaleRange()
 {
-  QFETCH( float, scale );
+  QFETCH( double, scale );
   QFETCH( bool, isInScale );
 
-  mpLayer->setMinimumScale( 2500.0f );
-  mpLayer->setMaximumScale( 5000.0f );
+  mpLayer->setMinimumScale( 2500.0 );
+  mpLayer->setMaximumScale( 5000.0 );
   mpLayer->setScaleBasedVisibility( true );
   QCOMPARE( mpLayer->isInScaleRange( scale ), isInScale );
   //always in scale range if scale based visibility is false
